+++ src/app/components/Cards/BoardCard/BoardCard.tsx
@@ -87,7 +87,11 @@ export default function BoardCard({
           </div>
           {member.affiliations.length > 0 && (
             <div className="boardCard-full-initial-affiliations">
-              <AffiliationLabel affiliation={member.affiliations[0]} />
+              {member.affiliations.map((affiliation, index) => (
+                <div key={index}>
+                  <AffiliationLabel affiliation={affiliation} />
+                </div>
+              ))}
             </div>
           )}
           {member.assignedSections.length > 0 && (
@@ -234,7 +238,11 @@ export default function BoardCard({
       </div>
       {member.affiliations.length > 0 && (
         <div className="boardCard-affiliations">
-          <AffiliationLabel affiliation={member.affiliations[0]} />
+          {member.affiliations.map((affiliation, index) => (
+            <div key={index}>
+              <AffiliationLabel affiliation={affiliation} />
+            </div>
+          ))}
         </div>
       )}
       {member.assignedSections.length > 0 && (